detect: move unittests into tests/
